Move TensorKind and Selectors to types.py
