+++ tiledb/ml/readers/pytorch.py
@@ -11,9 +11,9 @@ from torch.utils.data import DataLoader, IterDataPipe
 from torchdata.datapipes.iter import IterableWrapper
 
 from ._pytorch_collators import Collator
-from ._tensor_schema import TensorKind, TensorSchema
+from ._tensor_schema import TensorSchema
 from ._tensor_schema.ranges import InclusiveRange
-from .types import ArrayParams
+from .types import ArrayParams, TensorKind
 
 TensorLike = Union[np.ndarray, sparse.COO, scipy.sparse.csr_matrix]
 TensorLikeOrTuple = Union[TensorLike, Tuple[TensorLike, ...]]
