+++ tiledb/ml/readers/types.py
@@ -1,11 +1,24 @@
+import enum
 from dataclasses import dataclass, field
-from typing import Any, Mapping, Optional, Sequence, Union
+from typing import TYPE_CHECKING, Any, Mapping, Optional, Sequence, Union
 
 import numpy as np
 
 import tiledb
 
-from ._tensor_schema import Selector, TensorKind, TensorSchema, TensorSchemaFactories
+if TYPE_CHECKING:
+    from ._tensor_schema import TensorSchema
+
+Selector = Union[slice, Sequence[int]]
+
+
+class TensorKind(enum.Enum):
+    """Kind of tensor."""
+
+    DENSE = enum.auto()
+    SPARSE_COO = enum.auto()
+    SPARSE_CSR = enum.auto()
+    RAGGED = enum.auto()
 
 
 @dataclass(frozen=True)
@@ -29,7 +42,7 @@ class ArrayParams:
     fields: Sequence[str] = ()
     dim_selectors: Mapping[str, Selector] = field(default_factory=dict)
     tensor_kind: Optional[TensorKind] = None
-    tensor_schema: TensorSchema[Any] = field(init=False, repr=False)
+    _tensor_schema_kwargs: Mapping[str, Any] = field(init=False, repr=False)
 
     def __post_init__(self) -> None:
         all_attrs = [self.array.attr(i).name for i in range(self.array.nattr)]
@@ -86,8 +99,7 @@ class ArrayParams:
         else:
             tensor_kind = TensorKind.SPARSE_COO
 
-        factory = TensorSchemaFactories[sparse, tensor_kind]
-        tensor_schema = factory(
+        tensor_schema_kwargs = dict(
             kind=tensor_kind,
             _array=self.array,
             _key_dim_index=key_dim_index,
@@ -97,4 +109,13 @@ class ArrayParams:
             _dim_selectors=dim_selector_indices,
             _query_kwargs={"attrs": tuple(attrs), "dims": tuple(dims)},
         )
-        object.__setattr__(self, "tensor_schema", tensor_schema)
+        object.__setattr__(self, "_tensor_schema_kwargs", tensor_schema_kwargs)
+
+    @property
+    def tensor_schema(self) -> "TensorSchema[Any]":
+        # local import to avoid cyclic import
+        from ._tensor_schema import TensorSchemaFactories
+
+        kwargs = self._tensor_schema_kwargs
+        factory = TensorSchemaFactories[self.array.schema.sparse, kwargs["kind"]]
+        return factory(**kwargs)
